+++ 1-foundation/practical-logger-browser/src/sinks/common.ts
@@ -1,3 +1,4 @@
+import { LogLevel, LOG_LEVEL_TO_HUMAN } from '@offirmo/practical-logger-core'
 
 export const LEVEL_TO_CONSOLE_METHOD: Readonly<{ [k: string]: string }> = {
 	fatal:   'error',
@@ -19,4 +20,13 @@ export const LEVEL_TO_CONSOLE_METHOD: Readonly<{ [k: string]: string }> = {
 	debug:   'debug',
 	trace:   'debug',
 	silly:   'debug',
+}
+
+
+const MIN_WIDTH = 5
+export function to_uniform_level(level: LogLevel): string {
+	let str = LOG_LEVEL_TO_HUMAN[level] //.slice(0, MIN_WIDTH)
+	//if (str.length < MIN_WIDTH)
+	str = (str + '         ').slice(0, MIN_WIDTH)
+	return str
 }