+++ 1-foundation/practical-logger-browser/src/types.ts
@@ -1,5 +1,8 @@
 import { BaseSinkOptions } from '@offirmo/practical-logger-types'
 
+export type Browser = 'firefox' | 'safari' | 'chromium' | 'other'
+
 export interface SinkOptions extends BaseSinkOptions {
-	fancy?: boolean
+	useCss?: boolean
+	explicitBrowser?: Browser
 }